backend/requirements.txt
Django>=3.0
django-cors-headers>=3.2.0
djangorestframework>=3.10.3
psycopg2>=2.8.4
testfixtures>=6.10.3
